% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_monitoring_check.R
\name{data_monitoring_check}
\alias{data_monitoring_check}
\title{This function filters down data monitoring/inverter down and merges with reference files.}
\usage{
data_monitoring_check(se, sma, sp, status, clients)
}
\arguments{
\item{se}{SolarEdge Data.}

\item{sma}{SunnyPortal Data}

\item{sp}{SunPower Data.}

\item{clients}{Reference for customer job information.}

\item{staus}{Reference list on site status.}
}
\value{
Updated recent down data monitoring data.frame as an object, and also written to recent_down_dm.csv.
}
\description{
This function filters down data monitoring/inverter down and merges with reference files.
}
\examples{
data_monitoring_check<- function(se, sma, sp, status, clients)
}
