% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMDOWNCHECK.R
\name{DMDOWNCHECK}
\alias{DMDOWNCHECK}
\title{This function filters down data monitoring/inverter down and merges with reference files.}
\usage{
DMDOWNCHECK(SE, SMA, SP, RECENTDDMLIST, FINALPORTALKEY)
}
\arguments{
\item{SE}{SolarEdge Data.}

\item{SMA}{SunnyPortal Data}

\item{SP}{SunPower Data.}

\item{RECENTDDMLIST}{Reference list on site status.}

\item{FINALPORTALKEY}{Reference for customer job information.}
}
\value{
Updated RECENTDDMALIST.
}
\description{
This function filters down data monitoring/inverter down and merges with reference files.
}
\examples{
DMDOWNCHECK(SE, SMA, SP, RECENTDDMLIST, FINALPORTALKEY)
}
