% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly_above.R
\name{monthly_positive_deviation}
\alias{monthly_positive_deviation}
\title{This function compares estimated monthly yield to actual yield for portal SE and SMA.}
\usage{
monthly_positive_deviation(sma_performance, se_performance, estimates, status)
}
\arguments{
\item{sma_performance}{SunnyPortal Data}

\item{se_performance}{SolarEdge Data.}

\item{estimates}{Reference for .}

\item{status}{Reference list on site status.}
}
\value{
Monthly production above estimates data.frame as an object, and also written to monthly_positive_deviation.csv.
}
\description{
This function compares estimated monthly yield to actual yield for portal SE and SMA.
}
\examples{
monthly_positive_deviation(sma_performance,se_performance,estimates,status)
}
