% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MONTHLYDEVIATION.R
\name{MONTHYLDEVIATION}
\alias{MONTHYLDEVIATION}
\title{This function compares estimated monthly yield to actual yield for portal SE and SMA.}
\usage{
MONTHYLDEVIATION(SMA, SE, PERFORMANCEKEY, RECENTDDMLIST)
}
\arguments{
\item{SMA}{SunnyPortal Data}

\item{SE}{SolarEdge Data.}

\item{PERFORMANCEKEY}{Reference for .}

\item{RECENTDDMLIST}{Reference list on site status.}
}
\value{
Monthly yield deviations.
}
\description{
This function compares estimated monthly yield to actual yield for portal SE and SMA.
}
\examples{
MONTHLYDEVIATION(SE, SMA, PERFORMANCEKEY,RECENTDDMLIST)
}
