% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearly_above.R
\name{yearly_positive_deviation}
\alias{yearly_positive_deviation}
\title{This function compares estimated yearly yield to actual yield for portals SE and SMA.}
\usage{
yearly_positive_deviation(sma_performance, se_performance, estimates, status)
}
\arguments{
\item{sma_performance}{SunnyPortal Data}

\item{se_performance}{SolarEdge Data.}

\item{estimates}{Reference for .}

\item{status}{Reference list on site status.}
}
\value{
Yearly production above estimates data.frame as an object, and also written to yearly_positive_deviation.csv.
}
\description{
This function compares estimated yearly yield to actual yield for portals SE and SMA.
}
\examples{
yearly_positive_deviation(sma_performance,se_performance,estimates,status)
}
