% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YEARLYDEVIATION.R
\name{YEARLYDEVIATION}
\alias{YEARLYDEVIATION}
\title{This function compares estimated yearly yield to actual yield for portals SE and SMA.}
\usage{
YEARLYDEVIATION(SMA, SE, PERFORMANCEKEY, RECENTDDMLIS)
}
\arguments{
\item{SMA}{SunnyPortal Data}

\item{SE}{SolarEdge Data.}

\item{PERFORMANCEKEY}{Reference for .}

\item{RECENTDDMLIST}{Reference list on site status.}
}
\value{
Yearly yield deviations.
}
\description{
This function compares estimated yearly yield to actual yield for portals SE and SMA.
}
\examples{
YEARLYDEVIATION(SE, SMA, PERFORMANCEKEY, RECENTDDMLIST)
}
